requirements.txt
keras  == 2.12.0 
tensorflow == 2.12.0
numpy == 1.23.5
google-cloud-storage == 2.9.0
Flask == 2.3.2
pillow == 9.5.0
gunicorn == 20.1.0
python-dotenv == 1.0.0
# appengine-python-standard == 1.1.2